+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Version 0.6.0 April 2020
 - **(C)** Improved text cleaning to remove page builder tags
+- **(C)** Code updates and improvements
 
 ### Version 0.5.9 March 2020
 - **(C)** Improved error catching in Version Check form field